+++ asteroids/renderer.h
@@ -9,10 +9,10 @@
 #define SCREEN_HEIGHT 480
 
 //assing a colour to a single pixel
-int draw_pixel(uint32_t* pixel_buffer, int x, int y, uint32_t colour);
+int draw_pixel(uint16_t* pixel_buffer, int x, int y, uint16_t colour);
 
-int draw_line(uint32_t* pixel_buffer, int x1, int y1, int x2, int y2, uint32_t colour);
+int draw_line(uint16_t* pixel_buffer, int x1, int y1, int x2, int y2, uint16_t colour);
 
 //assign a colour to all pixels
-void clear_pixels(uint32_t* pixel_buffer, uint32_t colour);
+void clear_pixels(uint16_t* pixel_buffer, uint16_t colour);
